+++ packages/eip-5792-middleware/src/index.ts
@@ -1,11 +1,27 @@
 export type {
   ProcessSendCallsRequest,
   ProcessSendCallsHooks,
-} from './methods/processSendCalls';
-export { processSendCalls } from './methods/processSendCalls';
-export { getCallsStatus } from './methods/getCallsStatus';
+} from './hooks/processSendCalls';
+export { processSendCalls } from './hooks/processSendCalls';
+export { getCallsStatus } from './hooks/getCallsStatus';
 export {
   getCapabilities,
   type GetCapabilitiesHooks,
-} from './methods/getCapabilities';
+} from './hooks/getCapabilities';
+export { walletSendCalls } from './methods/wallet_sendCalls';
+export { walletGetCallsStatus } from './methods/wallet_getCallsStatus';
+export { walletGetCapabilities } from './methods/wallet_getCapabilities';
 export type { EIP5792Messenger } from './types';
+
+export type {
+  GetCallsStatusHook,
+  GetCallsStatusParams,
+  GetCallsStatusResult,
+  GetCapabilitiesHook,
+  GetCapabilitiesParams,
+  GetCapabilitiesResult,
+  ProcessSendCallsHook,
+  SendCallsPayload as SendCalls,
+  SendCallsParams,
+  SendCallsResult,
+} from './types';