+++ packages/transaction-controller/src/hooks/CollectPublishHook.test.ts
@@ -1,3 +1,5 @@
+import { noop } from 'lodash';
+
 import { CollectPublishHook } from './CollectPublishHook';
 import type { TransactionMeta } from '..';
 import { flushPromises } from '../../../../tests/helpers';
@@ -10,21 +12,43 @@ const ERROR_MESSAGE_MOCK = 'Test error';
 
 const TRANSACTION_META_MOCK = {
   id: '123-456',
+  txParams: {
+    nonce: '0x1',
+  },
+} as TransactionMeta;
+
+const TRANSACTION_META_2_MOCK = {
+  id: '123-457',
+  txParams: {
+    nonce: '0x2',
+  },
 } as TransactionMeta;
 
 describe('CollectPublishHook', () => {
   describe('getHook', () => {
-    it('returns function that resolves ready promise', async () => {
+    it('resolves ready promise', async () => {
       const collectHook = new CollectPublishHook(2);
       const publishHook = collectHook.getHook();
 
-      publishHook(TRANSACTION_META_MOCK, SIGNED_TX_MOCK).catch(() => {
-        // Intentionally empty
-      });
+      publishHook(TRANSACTION_META_MOCK, SIGNED_TX_MOCK).catch(noop);
+      publishHook(TRANSACTION_META_2_MOCK, SIGNED_TX_2_MOCK).catch(noop);
 
-      publishHook(TRANSACTION_META_MOCK, SIGNED_TX_2_MOCK).catch(() => {
-        // Intentionally empty
-      });
+      await flushPromises();
+
+      const result = await collectHook.ready();
+
+      expect(result.signedTransactions).toStrictEqual([
+        SIGNED_TX_MOCK,
+        SIGNED_TX_2_MOCK,
+      ]);
+    });
+
+    it('resolves ready promise with signatures in nonce order', async () => {
+      const collectHook = new CollectPublishHook(2);
+      const publishHook = collectHook.getHook();
+
+      publishHook(TRANSACTION_META_2_MOCK, SIGNED_TX_2_MOCK).catch(noop);
+      publishHook(TRANSACTION_META_MOCK, SIGNED_TX_MOCK).catch(noop);
 
       await flushPromises();
 
@@ -48,8 +72,31 @@ describe('CollectPublishHook', () => {
       );
 
       const publishPromise2 = publishHook(
-        TRANSACTION_META_MOCK,
+        TRANSACTION_META_2_MOCK,
         SIGNED_TX_2_MOCK,
+      );
+
+      collectHook.success([TRANSACTION_HASH_MOCK, TRANSACTION_HASH_2_MOCK]);
+
+      const result1 = await publishPromise1;
+      const result2 = await publishPromise2;
+
+      expect(result1.transactionHash).toBe(TRANSACTION_HASH_MOCK);
+      expect(result2.transactionHash).toBe(TRANSACTION_HASH_2_MOCK);
+    });
+
+    it('resolves all publish promises in nonce order', async () => {
+      const collectHook = new CollectPublishHook(2);
+      const publishHook = collectHook.getHook();
+
+      const publishPromise2 = publishHook(
+        TRANSACTION_META_2_MOCK,
+        SIGNED_TX_2_MOCK,
+      );
+
+      const publishPromise1 = publishHook(
+        TRANSACTION_META_MOCK,
+        SIGNED_TX_MOCK,
       );
 
       collectHook.success([TRANSACTION_HASH_MOCK, TRANSACTION_HASH_2_MOCK]);